Implement arithmetic.test_numeric and arithmetic.test_object (#22254)
